# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Planificacion Horaria",
        "description": "Gestion de horarios de clases",
        "contact": {
            "email": "contacto@tudominio.com"
        },
        "license": {
            "name": "Licencia",
            "url": "https://www.terciarioUrquiza.com.ar/planificacionHoraria"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/horarios": {
            "get": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Obtener todos los horarios",
                "description": "Retorna un array de horarios",
                "operationId": "d6e7d1b6cad06a6b841a864d9ffeaa2c",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "No se encontraron horarios"
                    },
                    "500": {
                        "description": "Error al obtener los horarios"
                    }
                }
            }
        },
        "/api/horarios/{id}": {
            "get": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Obtener horario por id",
                "description": "Retorna un horario",
                "operationId": "f3d0940996bf0313a64f7e9ea3021ea4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del horario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "No se encontró el horario"
                    },
                    "500": {
                        "description": "Error al obtener el horario"
                    }
                }
            }
        },
        "/api/horarios/guardar": {
            "post": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Guardar horario",
                "description": "Guardar un nuevo horario",
                "operationId": "f8714b29f02aebc34b464b3edc09fdfc",
                "requestBody": {
                    "description": "Datos del horario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Horario"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Horario guardado correctamente"
                    },
                    "500": {
                        "description": "Error al guardar el horario"
                    }
                }
            }
        },
        "/api/horarios/actualizar/{id}": {
            "put": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Actualizar horario",
                "description": "Actualizar un horario existente",
                "operationId": "714fa83f5ddd9e3ff9acecefd3c34100",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del horario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del horario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Horario"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Horario actualizado correctamente"
                    },
                    "404": {
                        "description": "No se encontró el horario"
                    },
                    "500": {
                        "description": "Error al actualizar el horario"
                    }
                }
            }
        },
        "/api/horarios/eliminar/{id}": {
            "delete": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Eliminar horario",
                "description": "Eliminar un horario existente",
                "operationId": "a465abe784ef0fc74755bfddbff1d392",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del horario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Horario eliminado correctamente"
                    },
                    "404": {
                        "description": "No se encontró el horario"
                    },
                    "500": {
                        "description": "Error al eliminar el horario"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Horario": {
                "title": "Horario",
                "description": "Esquema del objeto Horario",
                "properties": {
                    "id_horario": {
                        "description": "ID del horario",
                        "type": "integer"
                    },
                    "dia": {
                        "description": "Dia de la semana",
                        "type": "string"
                    },
                    "hora_inicio": {
                        "description": "Hora de inicio",
                        "type": "time"
                    },
                    "hora_fin": {
                        "description": "Hora de fin",
                        "type": "time"
                    },
                    "v_p": {
                        "description": "Vigente o Pasado",
                        "type": "boolean"
                    },
                    "id_dm": {
                        "description": "ID de la relacion docente materia",
                        "type": "integer"
                    },
                    "id_aula": {
                        "description": "ID del aula",
                        "type": "integer"
                    },
                    "id_comision": {
                        "description": "ID de la comision",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    }
}